# api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "."
  version: "1.0.0"
  title: "MongoLeafDataGuard API"
host: "localhost:3001"
basePath: "/dataguard/v1"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /users:
    get:
      tags:
      - "Users"
      summary: "Get users"
      operationId: "getUsers"
      parameters:
      - name: "user_id"
        in: "query"
        description: "?user_id="
        required: false
        type: "string"
      - in: "query"
        name: "limit"
        type: "number"
        required: false
        description: "Limit results (?limit=5&page=1)"
      - in: "query"
        name: "page"
        type: "number"
        required: false
        description: "Limit results (?limit=5&page=1)"
      responses:
        200:
          description: "Users"
          schema:
            $ref: "#/definitions/Dataresponse"
        404:
          description: "No users found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal error"
      x-swagger-router-controller: "Users"
    post:
      tags:
      - "Users"
      summary: "Create new user"
      operationId: "createUser"
      parameters:
      - in: "body"
        name: "user"
        description: "Username and userpassword"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        201:
          description: "User created"
      x-swagger-router-controller: "Users"
    delete:
      tags:
      - "Users"
      summary: "Deletes User"
      operationId: "deleteUser"
      parameters:
      - name: "user_id"
        in: "query"
        description: "?user_id=testuser"
        required: true
        type: "string"
      responses:
        200:
          description: "User deleted"
          schema:
            $ref: "#/definitions/Dataresponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal error"
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "Users"
  /auth:
    get:
      tags:
      - "Authentifications"
      summary: "verify user token"
      operationId: "authUser"
      parameters:
      - in: "query"
        name: "token"
        description: "?token="
        required: true
        type: "string"
      responses:
        201:
          description: "OK"
          schema:
            $ref: "#/definitions/Dataresponse"
        403:
          description: "Access denied"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal error"
      x-swagger-router-controller: "Authentifications"
    x-swagger-router-controller: "Authentifications"
...
definitions:
...
  Dataresponse:
    type: "object"
    required:
    - "data"
    - "message"
    properties:
      message:
        type: "string"
      data:
        type: "array"
        items:
          type: "string"
...
  User:
    type: "object"
    properties:
      username:
        type: "string"
...
  Error:
    type: object
    properties:
      error:
        type: integer
      message:
        type: string